+++ src/components/navbar/BurgerFancy/_BurgerFancy.scss
@@ -1,4 +1,4 @@
-@import "../../../../functionality/styler/_ThemedStyles.scss" ;
+@import "../../../functionality/styler/_ThemedStyles.scss";
 
 .Header__Burger {
     cursor: pointer;
