+++ src/pages/blog/blog.scss
@@ -84,8 +84,3 @@ body {
     margin-right: auto;
   }
 }
-
-.BlogContainer {
-  &__Header {
-  }
-}
